+++ utils/gnomad_functions.py
@@ -238,201 +238,3 @@ def melt_kt_grouped(kt, columns_to_melt, value_column_names, key_column_name='va
             .drop('comb'))
     """
     raise NotImplementedError
-
-
-def get_duplicated_samples(
-        kin_ht: hl.Table,
-        i_col: str = 'i',
-        j_col: str = 'j',
-        kin_col: str = 'kin',
-        duplicate_threshold: float = 0.4
-) -> List[Set[str]]:
-    """
-    Given a pc_relate output Table, extract the list of duplicate samples. Returns a list of set of samples that are duplicates.
-
-
-    :param Table kin_ht: pc_relate output table
-    :param str i_col: Column containing the 1st sample
-    :param str j_col: Column containing the 2nd sample
-    :param str kin_col: Column containing the kinship value
-    :param float duplicate_threshold: Kinship threshold to consider two samples duplicated
-    :return: List of samples that are duplicates
-    :rtype: list of set of str
-    """
-
-    def get_all_dups(s, dups, samples_duplicates):
-        if s in samples_duplicates:
-            dups.add(s)
-            s_dups = samples_duplicates.pop(s)
-            for s_dup in s_dups:
-                if s_dup not in dups:
-                    dups = get_all_dups(s_dup, dups, samples_duplicates)
-        return dups
-
-    dup_rows = kin_ht.filter(kin_ht[kin_col] > duplicate_threshold).collect()
-
-    samples_duplicates = defaultdict(set)
-    for row in dup_rows:
-        samples_duplicates[row[i_col]].add(row[j_col])
-        samples_duplicates[row[j_col]].add(row[i_col])
-
-    duplicated_samples = []
-    while len(samples_duplicates) > 0:
-        duplicated_samples.append(get_all_dups(list(samples_duplicates)[0], set(), samples_duplicates))
-
-    return duplicated_samples
-
-
-def infer_families(kin_ht: hl.Table,
-                   sex: Dict[str, bool],
-                   duplicated_samples: Set[str],
-                   i_col: str = 'i',
-                   j_col: str = 'j',
-                   kin_col: str = 'kin',
-                   ibd2_col: str = 'ibd2',
-                   first_degree_threshold: Tuple[float, float] = (0.2, 0.4),
-                   second_degree_threshold: Tuple[float, float] = (0.05, 0.16),
-                   ibd2_parent_offspring_threshold: float = 0.2
-                   ) -> hl.Pedigree:
-    """
-
-    Infers familial relationships from the results of pc_relate and sex information.
-    Note that both kinship and ibd2 are needed in the pc_relate output.
-
-    This function returns a pedigree containing trios inferred from the data. Family ID can be the same for multiple
-    trios if one or more members of the trios are related (e.g. sibs, multi-generational family). Trios are ordered by family ID.
-
-    Note that this function only returns complete trios defined as:
-    one child, one father and one mother (sex is required for both parents)
-
-    :param Table kin_ht: pc_relate output table
-    :param dict of str -> bool sex: A dict containing the sex for each sample. True = female, False = male, None = unknown
-    :param set of str duplicated_samples: Duplicated samples to remove (If not provided, this function won't work as it assumes that each child has exactly two parents)
-    :param str i_col: Column containing the 1st sample id in the pc_relate table
-    :param str j_col: Column containing the 2nd sample id in the pc_relate table
-    :param str kin_col: Column containing the kinship in the pc_relate table
-    :param str ibd2_col: Column containing ibd2 in the pc_relate table
-    :param (float, float) first_degree_threshold: Lower/upper bounds for kin for 1st degree relatives
-    :param (float, float) second_degree_threshold: Lower/upper bounds for kin for 2nd degree relatives
-    :param float ibd2_parent_offspring_threshold: Upper bound on ibd2 for a parent/offspring
-    :return: Pedigree containing all trios in the data
-    :rtype: Pedigree
-    """
-
-    def get_fam_samples(sample: str,
-                        fam: Set[str],
-                        samples_rel: Dict[str, Set[str]],
-                        ) -> Set[str]:
-        """
-        Given a sample, its known family and a dict that links samples with their relatives, outputs the set of
-        samples that constitute this sample family.
-
-        :param str sample: sample
-        :param dict of str -> set of str samples_rel: dict(sample -> set(sample_relatives))
-        :param set of str fam: sample known family
-        :return: Family including the sample
-        :rtype: set of str
-        """
-        fam.add(sample)
-        for s2 in samples_rel[sample]:
-            if s2 not in fam:
-                fam = get_fam_samples(s2, fam, samples_rel)
-        return fam
-
-    def get_indexed_ibd2(
-            pc_relate_rows: List[hl.Struct]
-    ) -> Dict[Tuple[str, str], float]:
-        """
-        Given rows from a pc_relate table, creates a dict with:
-        keys: Pairs of individuals, lexically ordered
-        values: ibd2
-
-        :param list of hl.Struct pc_relate_rows: Rows from a pc_relate table
-        :return: Dict of lexically ordered pairs of individuals -> kinship
-        :rtype: dict of (str, str) -> float
-        """
-        ibd2 = dict()
-        for row in pc_relate_rows:
-            ibd2[tuple(sorted((row[i_col], row[j_col])))] = row[ibd2_col]
-        return ibd2
-
-    def get_parents(
-            possible_parents: List[str],
-            indexed_kinship: Dict[Tuple[str, str], Tuple[float, float]],
-            sex: Dict[str, bool]
-    ) -> Tuple[str, str]:
-        """
-        Given a list of possible parents for a sample (first degree relatives with low ibd2),
-        looks for a single pair of samples that are unrelated with different sexes.
-        If a single pair is found, return the pair (father, mother)
-
-        :param list of str possible_parents: Possible parents
-        :param dict of (str, str) -> (float, float)) indexed_kinship: Dict mapping pairs of individuals to their kinship and ibd2 coefficients
-        :param dict of str -> bool sex: Dict mapping samples to their sex (True = female, False = male, None or missing = unknown)
-        :return: (father, mother)
-        :rtype: (str, str)
-        """
-
-        parents = []
-        while len(possible_parents) > 1:
-            p1 = possible_parents.pop()
-            for p2 in possible_parents:
-                if tuple(sorted((p1,p2))) not in indexed_kinship:
-                    if sex.get(p1) is False and sex.get(p2):
-                        parents.append((p1,p2))
-                    elif sex.get(p1) and sex.get(p2) is False:
-                        parents.append((p2,p1))
-
-        if len(parents) == 1:
-            return parents[0]
-
-        return None
-
-    # Get first degree relatives - exclude duplicate samples
-    dups = hl.literal(duplicated_samples)
-    first_degree_pairs = kin_ht.filter(
-        (kin_ht[kin_col] > first_degree_threshold[0]) &
-        (kin_ht[kin_col] < first_degree_threshold[1]) &
-        ~dups.contains(kin_ht[i_col]) &
-        ~dups.contains(kin_ht[j_col])
-    ).collect()
-    first_degree_relatives = defaultdict(set)
-    for row in first_degree_pairs:
-        first_degree_relatives[row[i_col]].add(row[j_col])
-        first_degree_relatives[row[j_col]].add(row[i_col])
-
-    # Add second degree relatives for those samples
-    # This is needed to distinguish grandparent - child - parent from child - mother, father down the line
-    first_degree_samples = hl.literal(set(first_degree_relatives.keys()))
-    second_degree_samples = kin_ht.filter(
-        (first_degree_samples.contains(kin_ht[i_col]) | first_degree_samples.contains(kin_ht[j_col])) &
-        (kin_ht[kin_col] > second_degree_threshold[0]) &
-        (kin_ht[kin_col] < first_degree_threshold[1])
-    ).collect()
-
-    ibd2 = get_indexed_ibd2(second_degree_samples)
-
-    fam_id = 1
-    trios = []
-    while len(first_degree_relatives) > 0:
-        s_fam = get_fam_samples(list(first_degree_relatives)[0], set(),
-                                first_degree_relatives)
-        for s in s_fam:
-            s_rel = first_degree_relatives.pop(s)
-            possible_parents = []
-            for rel in s_rel:
-                if ibd2[tuple(sorted((s, rel)))] < ibd2_parent_offspring_threshold:
-                    possible_parents.append(rel)
-
-            parents = get_parents(possible_parents, ibd2, sex)
-
-            if parents is not None:
-                trios.append(hl.Trio(s=s,
-                                     fam_id=str(fam_id),
-                                     pat_id=parents[0],
-                                     mat_id=parents[1],
-                                     is_female=sex.get(s)))
-
-        fam_id += 1
-
-    return hl.Pedigree(trios)
